Move lighting into its own file
